Step 5 of "Layer"ization - more robust.
IMPORTANT LESSON:
In python, If a list is a member of a base class OR a list is a member of a class whose instance is a member of a base class; then it must be initialized in the constructor of its immediately enclosing base class. Otherwise it gets shared by multiple instances of the top level class.