this makes cloud calculation as before. Yet the 'layer'rization has achieved better control over cloud overlay during the dynamic widgetQ manipulations.
This will also reduce wastefulness of cloud calculation - because it is done at the App level and not during widgetQ operations. This allows one time cloud calculations after a batch of widgetQ manipulations has been done. This wasn't intentional, but a fringe benefit. Some trick will have to be made to hide explicit calls to cloud calculation, so that the app programmer doesn't have to bother about it. But that's not a priority though.