[Changelog] inkface.loadsvg() loads the SVG elements in a dict object rather than a list.
The dict is keyed on element's 'name' attribute. This will improve the application code both in performance and readability - because now there is no need to iterate over entire list and referring to elements by their 'name' is more intuitive.

The tests are slightly changed to work with this change, however they will be further modified to be more readable as said above.

The canvas still registers the stored elements as list, hence passing them as a list to event handlers. This will be changed too in future changeset.