This changeset makes the "paint" callback function a characteristic of X canvas, instead of a generic one. OpenGL might have other kind of paint mechanism.
The paint handler that python wrapper registers with the cobject, is now not passed as part of canvas_init, but there is a separate init function for it.

The liniking to OpenGL library has been commented out for this checkin, because it's causing crashes during exit of app. This changeset is not enabling any OpenGL functionality anyway, this is just a reorg step towards that goal.