[Changelog] Introduction of cloud algorithm to correctly implement event generation for overlapping elements.
This introduces the cloud algorithm. The elements currently present in the canvas list are assumed to be in the same order as they are displayed. The cloud algorithm, calculates the overlapping rectangles between two elements and stores them as clouds in the lower elements. During mouse pointer processing, if the coordinates of the pointer event are found to come over these clouds then that event is ignored for the lower element. (The later logic is not part of this changeset though)