This provides a new interface that helps in reducing CPU usage drastically.
For animation we fork a PainterThread which redraws the canvas FRAMERATE number of times in 1 sec. - even when we are not animating. This is wasteful, because unlike games (main target of pygame), UI doesn't need constant movements of the interface. So an interface method was added to canvas object which will let the app decide "at what frame rate" and "for how many frames" will the animation will progress. This makes the PainterThread to invoke canvas painting only when it is needed.