Improving API needed for animation and the performance of animation [1/2]
This change to some extent undoes some changes done in r994.

With this change, the canvas uses pygame's "LayeredDirty" kind of sprite groups.
The draw method of this kind of group repaints the sprites only when they are
marked dirty. Using this feature now, we can run the painter thread continuously
and use it only when there are dirty sprites/elements. It was verified that this
approach works in reducing the CPU usage to far low levels ( < 25% during active
animation and < 5% during no changes to GUI - as opposed to constant ~40%+ CPU
usage in past). 

Also due to this change, app programmers don't have to call the "animate"
function explicitly. 

To find out the times when element surface get dirty - few new API method were
added to the element. Element surfaces get dirty (need repaint) only when their
position is changed OR their SVG composition is changed. The later is already
done by element's "refresh" method. To capture the former case, a new API was
added to Element "set_position". Now CanvasElement's x and y coordinates
are private variables and can only be accessed via {get,set}_position methods.
During refresh and set_position, the dirty flag is set on the sprite of the
element.