This makes sure that by specifying framerate = 0 during canvas creation, the app programmer can tell the canvas that it's not interested in animation and canvas need not spawn a thread which will periodically keep updating the display.
In such case however it's app programmer's responsibility to call
canvas.update() everytime it wants the canvas to be repainted. That is typically
when the app changes some elements of GUI.

Several tests that don't use animation, have been modified to use framerate = 0,
hence saving from spawing an extra thread.