This is a solution to a problem.
Problem:
When a face is added to a canvas, its elements are copied into the queue maintained by the canvas. If the face's elements are cloned after it was added to the canvas, then the new copies of the elements do not show up on canvas. 

Solution:
For the new cloned elements to show up on canvas they have to be registered at two levels of canvas - the base 'Canvas' class and 'PygameCanvas' class. To achive this with minimum/no hassle to app programmer, this was automatic update was made part of the clone process.
The clone call on Face (and now PygameFace too) are used to update the Canvas's (and PygameCanvas's) queues. The new element that is result of the clone call, is inserted into the respective queues as a neighbor of their original copy.