A widget library for inkface apps.
These widgets are independent of backends (except for some corner cases, to be fixed). These widgets will help improve the speed of developing an app, even though the GUI is made out of arbitrary shapes rendered from SVG document. 

Any widget of this library can be instantiated, by giving it a set of elements that you load from your app's SVG file. These elements can be passed individually (like in case of textbox widget) or the whole face containing these elements can be passed (like in case of keyboard widget) to the constructor of the widget.

Corresponding to each widget, there will be test cases which will show how to use those widgets.